requirements.txt
# -*- coding: utf-8 -*-
# Dependências principais
selenium==4.27.1
webdriver-manager==4.0.2
python-dotenv==1.0.1
pywin32==308
requests==2.32.3

# Para processamento de dados
pandas==2.2.3
openpyxl==3.1.5
tabulate==0.9.0

# Para monitoramento de arquivos
watchdog==4.0.0

# Para logs e tempo
python-dateutil==2.8.2

# Para empacotamento (opcional)
pyinstaller==6.11.1

# Para desenvolvimento e testes (opcional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.6.1
